--- src/App.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Album, StickyFooter } from '@/lib';
-import { CustomModalDemo, ColorPalette } from '@/examples';
-import './App.css';
-
-const App = () => (
-  <div>
-    <Album
-      Heading="Photo Album"
-      SubContent="Something short and leading about the collection below—its contents, the creator, etc.
-          Make it short and sweet, but not too short so folks don't simply skip over it
-          entirely."
-      cameraIconClasses="cameraIconClasses"
-    />
-    <CustomModalDemo />
-    <StickyFooter />
-    <ColorPalette />
-  </div>
-);
-
-export default App;
