--- src/codeBox/index.css
@@ -1,6 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    padding-top: 50px;
-}
