--- src/codeBox/index.js
@@ -1,7 +1,0 @@
-import CodeSnippet from './CodeSnippet';
-import customCardCode from './customCradCode';
-import customModalCode from './customModalCode';
-import signInCode from './signInCode';
-import stickyFooterCode from './stickyFooterCode';
-
-export {CodeSnippet, customCardCode, customModalCode, signInCode, stickyFooterCode};
