+++ src/components/codeBox/index.css
@@ -1,0 +1,6 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  padding-top: 50px;
+}
