+++ src/components/codeBox/signInCode.js
@@ -1,0 +1,14 @@
+const signInCode = `import React from 'react';
+import {SignIn} from "../node_modules/designops/dist/index.js";
+
+const SignInDemo = () => {
+    return(
+        <SignIn firstLabel={'User Name'}/>
+    ) 
+
+}
+
+export default SignInDemo;
+
+`;
+export default signInCode;
