+++ src/constants/Constants.js
@@ -1,0 +1,2 @@
+export const IMAGE =
+  'https://images.unsplash.com/photo-1617665146086-14d150bd34fc?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&amp;ixlib=rb-1.2.1&amp;auto=format&amp;fit=crop&amp;w=1000&amp;q=80';
