--- src/examples/CustomCardDemo/index.css
@@ -1,11 +1,0 @@
-.cardClass {
-    height: 30%;
-    width: 30%;
-    margin: 10px;
-  }
-  
-.container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
