--- src/examples/CustomModalDemo/index.css
@@ -1,3 +1,0 @@
-.buttonCss{
-    color:black !important;
-}
