--- src/examples/index.js
@@ -1,7 +1,0 @@
-import CustomModalDemo from './CustomModalDemo';
-import CustomCardDemo from './CustomCardDemo';
-import SignInDemo from './SignInDemo';
-import StickyFooterDemo from './StickyFooterDemo';
-import ColorPalette from './ColorPalette';
-import FontSizeDemo from './FontSizeDemo'
-export { CustomModalDemo, CustomCardDemo, SignInDemo, StickyFooterDemo, ColorPalette, FontSizeDemo};
