+++ src/pages/ColorPalette/index.js
