+++ src/pages/CustomModalDemo/index.css
@@ -1,0 +1,3 @@
+.buttonCss {
+  color: black !important;
+}
