--- src/routes.js
@@ -1,30 +1,0 @@
-
-
- const ROUTES=[
-    {
-        path:'/',
-        name:'Color Palette',
-    },
-    {
-        path:'/fontsize',
-        name:'Font Sizes',
-    },
-    {
-        path:'/signin',
-        name:'SignIn Page',
-    },
-    {
-        path:'/stickyfooter',
-        name:'Sticky Footer',
-    },
-    {
-        path:'/modal',
-        name:'Modal',
-    },
-    {
-        path:'/card',
-        name:'Cards',
-    }
-]
-
-export default ROUTES
